1.0.0 alpha提交 (#16)
优化以及支持功能:
需求缺陷管理
服务API管理
流水线管理
用例测试管理
用户组织、角色权限管理

+++ windy-client/src/main/java/com/zj/client/config/GlobalEnvConfig.java
@@ -22,7 +22,7 @@ public class GlobalEnvConfig {
   
   @Getter
   @Value("${windy.loop.query.timeout:7200000}")
-  private Integer loopQueryTimeout;
+  private Long loopQueryTimeout;
 
   @Setter
   @Getter
